+++ src/client.rs
@@ -1,0 +1,28 @@
+/// Common trait for swagger based client middleware
+pub trait Service {
+    /// Request body taken by client. Should implement `hyper::body::Payload`.
+    /// Likely either `hyper::Body`, `hyper::Chunk` or `swagger::ContextualPayload`.
+    type ReqBody;
+
+    /// Future response from client service.
+    /// Likely: `Future<Item=hyper::Response<hyper::Body>, Error=hyper::Error>`
+    type Future;
+
+    /// Handle the given request
+    fn request(&self, req: hyper::Request<Self::ReqBody>) -> Self::Future;
+}
+
+impl<C,B> Service for hyper::Client<C, B> where
+    B: hyper::body::Payload + Send + 'static,
+    B::Data: Send,
+    C: hyper::client::connect::Connect + Sync + 'static,
+    C::Transport: 'static,
+    C::Future: 'static,
+{
+    type ReqBody = B;
+    type Future = hyper::client::ResponseFuture;
+
+    fn request(&self, req: hyper::Request<Self::ReqBody>) -> Self::Future {
+        hyper::Client::request(self, req)
+    }
+}
